Migrated more defaults over into marin defaults.
