+++ experiments/dpo_ultrafeedback.py
@@ -7,7 +7,7 @@ Run DPO on the Ultrafeedback preference dataset using Marin's executor framework
 
 from fray.cluster import ResourceConfig
 from levanter.data.text import PreferenceChatLmDatasetFormat
-from marin.defaults import SimpleDPOConfig, default_dpo
+from marin.defaults import SimpleDPOConfig, default_dpo, default_tokenize
 from marin.execution.executor import executor_main
 from marin.execution.types import output_path_of
 from marin.processing.tokenize import lm_data_config
@@ -16,7 +16,6 @@ from experiments.llama import LLAMA3_CHAT_STOP_TOKEN_IDS, llama_8b
 from experiments.marin_models import marin_tokenizer
 from experiments.models import llama_3_1_8b
 from experiments.posttrain.preference_datasets import get_preference_dataset
-from experiments.tokenization import default_tokenize
 
 DATASET_NAME = "HuggingFaceH4/ultrafeedback_binarized"
