+++ experiments/exp1880_sft_baseline.py
@@ -16,7 +16,7 @@ import re
 
 from fray.cluster import ResourceConfig
 from levanter.data.text import DEFAULT_LM_DATA_SHUFFLE, ChatLmDatasetFormat
-from marin.defaults import SimpleSFTConfig, default_sft
+from marin.defaults import SimpleSFTConfig, default_sft, default_tokenize
 from marin.execution.executor import executor_main
 from marin.execution.types import ExecutorStep
 from marin.processing.tokenize import lm_mixture_data_config
@@ -30,7 +30,6 @@ from experiments.posttrain.instruction_datasets import (
     SMOLTALK2_SPLITS,
     get_instruction_dataset,
 )
-from experiments.tokenization import default_tokenize
 
 SLUGIFY_PATTERN = re.compile(r"[^a-z0-9]+")
