+++ experiments/pretraining_datasets/dclm.py
@@ -4,12 +4,12 @@
 import dataclasses
 
 from levanter.data.text import TextLmDatasetFormat
+from marin.defaults import default_tokenize
 from marin.execution.executor import executor_main
 from marin.processing.tokenize import lm_mixture_data_config
 
 from experiments.llama import llama3_tokenizer
 from experiments.pretraining_datasets.simple import downloads, tokenized
-from experiments.tokenization import default_tokenize
 
 DCLM_MIXTURE_WEIGHTS = {
     # token counts are for neox tokenizer
