+++ experiments/tutorials/train_tiny_sweep_dclm_tpu.py
@@ -8,11 +8,11 @@ on a 30M parameter DCLM model using TPU hardware.
 import dataclasses
 
 from fray.cluster import ResourceConfig
+from marin.defaults import CORE_TASKS
 from marin.execution.executor import executor_main
 from marin.execution.types import versioned
 
 from experiments.defaults import default_train
-from experiments.evals.task_configs import CORE_TASKS
 from experiments.llama import llama_30m
 from experiments.pretraining_datasets.dclm import dclm_mixture_config_llama3
 from experiments.simple_train_config import SimpleTrainConfig
