+++ experiments/tutorials/train_tiny_sweep_tpu.py
@@ -23,13 +23,13 @@ from fray import client as fray_client
 from fray.cluster import ResourceConfig
 from fray.types import Entrypoint, JobRequest, create_environment
 from levanter.main.train_lm import TrainLmConfig
+from marin.defaults import CORE_TASKS
 from marin.execution.sweep import SweepTarget, claim_and_run
 from marin.execution.types import versioned
 from marin.training.run_environment import extras_for_resources
 from marin.training.training import resolve_training_env
 
 from experiments.defaults import _run_training_on_worker, prepare_lm_train
-from experiments.evals.task_configs import CORE_TASKS
 from experiments.llama import llama_30m
 from experiments.pretraining_datasets.simple import tokenized
 from experiments.simple_train_config import SimpleTrainConfig
