+++ experiments/two_stage/data.py
@@ -2,13 +2,13 @@
 # SPDX-License-Identifier: Apache-2.0
 
 from marin.datakit.download.huggingface import DownloadConfig, download_hf
+from marin.defaults import default_tokenize
 from marin.execution.types import ExecutorStep, this_output_path
 
 from experiments.llama import llama3_tokenizer
 from experiments.midtraining_datasets import finemath_3_plus_tokenized
 from experiments.pretraining_datasets import tokenize_dolma
 from experiments.pretraining_datasets.simple import tokenized
-from experiments.tokenization import default_tokenize
 
 dolma_components = tokenize_dolma()
